requirements.txt
flask
flask-cors
python-dotenv
requests
snowflake-connector-python
gunicorn
